+++ puma/utils/histogram.py
@@ -1,6 +1,8 @@
 """Helper function for histogram handling."""
 import numpy as np
 
+from puma.utils import logger
+
 
 def save_divide(
     numerator,
@@ -49,12 +51,13 @@ def save_divide(
     return ratio
 
 
-def hist_w_unc(
+def hist_w_unc(  # pylint: disable=too-many-arguments,too-many-locals
     arr,
     bins,
     bins_range=None,
     normed: bool = True,
-    weights=None,
+    weights: np.ndarray = None,
+    underoverflow: bool = False,
 ):
     """
     Computes histogram and the associated statistical uncertainty.
@@ -77,6 +80,8 @@ def hist_w_unc(
         to each entry. The uncertainty of bins with weighted entries is
         sqrt(sum_i{w_i^2}) where w_i are the weights of the entries in this bin.
         By default None.
+    underoverflow : bool, optional
+        Option to include under- and overflow values in outermost bins.
 
     Returns
     -------
@@ -95,6 +100,19 @@ def hist_w_unc(
     if weights is None:
         weights = np.ones(len(arr))
 
+    # Check if there are nan values in the input values
+    nan_mask = np.isnan(arr)
+    if np.sum(nan_mask) > 0:
+        logger.warning("Histogram values contain %i nan values!", np.sum(nan_mask))
+        # Remove nan values
+        arr = arr[~nan_mask]
+        weights = weights[~nan_mask]
+
+    # Check if there are inf values in the input values
+    inf_mask = np.isinf(arr)
+    if np.sum(inf_mask) > 0:
+        logger.warning("Histogram values contain %i +-inf values!", np.sum(inf_mask))
+
     # Calculate the counts and the bin edges
     counts, bin_edges = np.histogram(arr, bins=bins, range=bins_range, weights=weights)
 
@@ -103,6 +121,29 @@ def hist_w_unc(
     unc = np.sqrt(
         np.histogram(arr, bins=bins, range=bins_range, weights=weights**2)[0]
     )
+
+    if underoverflow:
+        # add two dummy bins (from outermost bins to +-infinity)
+        bins_with_overunderflow = np.hstack(
+            [np.array([-np.inf]), bin_edges, np.array([np.inf])]
+        )
+        # recalculate the histogram with this adjusted binning
+        counts, _ = np.histogram(arr, bins=bins_with_overunderflow, weights=weights)
+        counts[1] += counts[0]  # add underflow values to underflow bin
+        counts[-2] += counts[-1]  # add overflow values to overflow bin
+        counts = counts[1:-1]  # remove dummy bins
+
+        # calculate the sum of squared weights
+        sum_squared_weights = np.histogram(
+            arr, bins=bins_with_overunderflow, weights=weights**2
+        )[0]
+        # add sum of squared weights from under/overflow values to under/overflow bin
+        sum_squared_weights[1] += sum_squared_weights[0]
+        sum_squared_weights[-2] += sum_squared_weights[-1]
+        sum_squared_weights = sum_squared_weights[1:-1]  # remove dummy bins
+
+        unc = np.sqrt(sum_squared_weights)  # uncertainty is sqrt(sum_squared_weights)
+
     if normed:
         sum_of_weights = float(np.sum(weights))
         counts = save_divide(counts, sum_of_weights, 0)