+++ changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ### [Latest]
+- Fix handling of nan values in histograms [#125](https://github.com/umami-hep/puma/pull/125)
+
+- Adding support for under- and overflow bins in `puma.HistogramPlot` [#124](https://github.com/umami-hep/puma/pull/124)
 
 ### [v0.1.8] (2022/08/30)